implement serviceimpl

+++ medily-backend/src/main/java/com/medily/backend/service/custom/ActivityLogService.java
@@ -6,5 +6,5 @@ import java.util.List;
 
 public interface ActivityLogService {
     List<ActivityLogResponseDTO> getRecentLogs();
-    void log(Long userId, String action);
+    void log(Integer userId, String action);
 }
